third_party/WebKit/LayoutTests/editing/inserting/insert-div-019.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<p>foo</p>',
            '|<br>',
            '<p>bar</p>',
        '</div>',
    ].join(''),
    'insertParagraph',
    [
        '<div contenteditable>',
            '<p>foo</p>',
            '<div><br></div>',
            '|<br>',
            '<p>bar</p>',
        '</div>',
    ].join('')),
    'insert DIV+BR before BR');
</script>
